Move holders from util/ to model/ Add @ColumnInfo annotations
